requirements.txt
flask==2.3.3
numpy==1.24.3
pillow==10.0.0
tenseal==0.3.12
scikit-learn==1.3.0
matplotlib==3.7.2
werkzeug==2.3.7